Display wind speed in miles per hour
Honestly, this isn't particularly useful because the weather stations do
not appear to have even the faintest idea what the real wind speed is in,
say, Amarillo, Texas; this program reports two miles an hour, which I know
is probably off by an order of magnitude. I imagine that the stations are
located in sheltered areas or something like that.